+++ src/main/java/com/example/parking/domain/admin/user/service/AdminUserService.java
@@ -1,12 +1,15 @@
 package com.example.parking.domain.admin.user.service;
 
 import com.example.parking.domain.admin.user.dto.AdminUserResDto;
+import com.example.parking.domain.user.entity.UserRole;
 import com.example.parking.domain.user.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
+import com.example.parking.domain.user.entity.UserRole;
+
 
 @Service
 @RequiredArgsConstructor
@@ -18,12 +21,14 @@ public class AdminUserService {
 
     public Page<AdminUserResDto> getAdminUsers(String keyword, Pageable pageable) {
         if (keyword == null || keyword.isBlank()) {
-            return userRepository.findAll(pageable)
+            return userRepository.findByRole(UserRole.USER, pageable)
                     .map(AdminUserResDto::from);
         }
 
-        return userRepository.findByNameContainingIgnoreCaseOrEmailContainingIgnoreCase(
+        return userRepository.findByRoleAndNameContainingIgnoreCaseOrRoleAndEmailContainingIgnoreCase(
+                        UserRole.USER,
                         keyword,
+                        UserRole.USER,
                         keyword,
                         pageable
                 )
