+++ src/main/java/com/example/parking/domain/user/controller/UserController.java
@@ -1,6 +1,7 @@
 package com.example.parking.domain.user.controller;
 
 import com.example.parking.domain.user.dto.*;
+import com.example.parking.domain.user.service.AuthService;
 import com.example.parking.domain.user.service.UserService;
 import com.example.parking.global.response.RsData;
 import com.example.parking.global.security.CustomUserDetails;
@@ -15,6 +16,7 @@ import org.springframework.web.bind.annotation.*;
 public class UserController {
 
     private final UserService userService;
+    private final AuthService authService;
 
     // [CUS-06] 이메일 중복 체크 - 회원가입 시 이메일 중복 여부를 확인하는 API
     @GetMapping("/api/users/check-email")
@@ -34,7 +36,7 @@ public class UserController {
     // [CUS-08] 로그인 - 로그인 요청을 받아 사용자 인증 후 JWT 토큰 발급
     @PostMapping("/api/users/login")
     public ResponseEntity<RsData<LoginResDto>> login(@Valid @RequestBody LoginReqDto reqDto) {
-        LoginResDto data = userService.login(reqDto);
+        LoginResDto data = authService.login(reqDto);
         RsData<LoginResDto> rsData = new RsData<>("로그인이 완료되었습니다.", "200-2", data);
         return ResponseEntity.ok(rsData);
     }
@@ -42,7 +44,7 @@ public class UserController {
     // access token 만료 후 refresh token으로 새 access token을 발급받는 API
     @PostMapping("/api/users/refresh")
     public ResponseEntity<RsData<LoginResDto>> refresh(@Valid @RequestBody RefreshTokenReqDto reqDto) {
-        LoginResDto data = userService.refresh(reqDto);
+        LoginResDto data = authService.refresh(reqDto);
         RsData<LoginResDto> rsData = new RsData<>("토큰 재발급이 완료되었습니다.", "200-3", data);
         return ResponseEntity.ok(rsData);
     }
@@ -84,7 +86,7 @@ public class UserController {
     public ResponseEntity<RsData<Void>> logout(
             @AuthenticationPrincipal CustomUserDetails userDetails
     ) {
-        userService.logout(userDetails.getUserId());
+        authService.logout(userDetails.getUserId());
         RsData<Void> rsData = new RsData<>("로그아웃이 완료되었습니다.", "200-7");
         return ResponseEntity.ok(rsData);
     }
