+++ src/test/java/com/example/parking/domain/admin/reservation/service/AdminReservationServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.admin.reservation.service;
+
+public class AdminReservationServiceTest {
+}
