+++ src/test/java/com/example/parking/domain/admin/user/service/AdminUserServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.admin.user.service;
+
+public class AdminUserServiceTest {
+}
