+++ src/test/java/com/example/parking/domain/parkingLot/service/ParkingLotServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.parkingLot.service;
+
+public class ParkingLotServiceTest {
+}
