+++ src/test/java/com/example/parking/domain/parkingLot/service/ParkingLotSyncServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.parkingLot.service;
+
+public class ParkingLotSyncServiceTest {
+}
