+++ src/test/java/com/example/parking/domain/parkingspot/service/ParkingSpotServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.parkingspot.service;
+
+public class ParkingSpotServiceTest {
+}
