+++ src/test/java/com/example/parking/domain/payment/service/PaymentServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.payment.service;
+
+public class PaymentServiceTest {
+}
