+++ src/test/java/com/example/parking/domain/user/service/AuthServiceTest.java
@@ -1,0 +1,4 @@
+package com.example.parking.domain.user.service;
+
+public class AuthServiceTest {
+}
